Audio device controller implementation
This controller will hold the full audio device list and allow the user to enable/disable them or rename them.
The settings will be saved to persistent storage.

The intended usage is:
 -  The preference application will load the state from persistent storage on startup. When the user makes changes, they will be saved to persistent storage and the helper application will be notified.
 - The helper application loads the state from persistent storage on startup, and will reload it when notified that something changed.

Given that only one process can modify persistent storage, the controller will not perform state reconciliation; on loading, the previous state will be erased.