Helper app start support
The helper app manager can now start the helper app if it is not currently running, as long as the helper application bundle is in the bundle of the parent application.
Unfortunately, since the application is sandboxed, it can’t directly ask the helper to quit. This will be implemented in a later commit.