Implement the AudioSwitch preferences app
This was originally supposed to be a preference pane, but that required a more complicated design and some restrictions.

The preference pane specific code was moved to the prefpane-experiment branch.

This commit also changes AudioSwitchHelper’s AppGroup, removing the “Dev” suffix. That makes it able to communicate with the preferences app.