Add support for macOS 13 Ventura
macOS 13 adds a new interface to manage login items (that gives more control to
the user). Use it instead of the old API.

The new API also allows the app itself to be added to the login items; this
means that instead of having a main app and a helper app, everything can be done
in a single app.

This will probably be used in the next release since it greatly simplifies the
application. The only downside is that the application will need to be manually
enabled to be lauched at startup on macOS < 13.